fix: Header data fetch
