+++ src/components/Keyword.tsx
@@ -1,7 +1,6 @@
 import Link from 'next/link'
 import React from 'react'
 import style from './Keyword.module.scss'
-import { notoSans } from '@/app/layout'
 
 type Props = {
   id: string
